port Automate vgpu vm upgrade test from cnv-tests repo
