+++ tests/virt/node/gpu/vgpu/test_windows_vm_with_vgpu.py
@@ -11,12 +11,14 @@ from pytest_testconfig import config as py_config
 
 from tests.virt.node.gpu.constants import VGPU_DEVICE_NAME_STR, VGPU_PRETTY_NAME_STR
 from tests.virt.node.gpu.utils import (
-    get_gpu_device_name_from_windows_vm,
     install_nvidia_drivers_on_windows_vm,
     restart_and_check_gpu_exists,
+)
+from tests.virt.utils import (
+    get_gpu_device_name_from_windows_vm,
+    validate_pause_optional_migrate_unpause_windows_vm,
     verify_gpu_device_exists_in_vm,
 )
-from tests.virt.utils import validate_pause_optional_migrate_unpause_windows_vm
 from utilities.constants import Images
 from utilities.virt import (
     VirtualMachineForTestsFromTemplate,
